requirements.txt
contourpy==1.3.3
CTkColorPicker==0.9.0
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
fonttools==4.59.2
kiwisolver==1.4.9
matplotlib==3.10.6
mpmath==1.3.0
numpy==2.3.2
packaging==25.0
pillow==11.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
six==1.17.0
sympy==1.14.0
